--- src/main/java/com/uncleray/controller/PersonController.java
@@ -1,22 +1,0 @@
-package com.uncleray.controller;
-
-import com.uncleray.model.pojo.Person;
-import com.uncleray.service.PersonService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-public class PersonController {
-
-    @Autowired
-    PersonService service;
-
-    @GetMapping("/persons")
-    public List<Person> getPersons() {
-        return service.getPersonList();
-    }
-
-}
