--- src/main/java/com/uncleray/dao/dao/UserDao.java
@@ -1,6 +1,0 @@
-package com.uncleray.dao.dao;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserDao {
-}
