--- src/main/java/com/uncleray/dao/mapper/StudentMapper.java
@@ -1,10 +1,0 @@
-package com.uncleray.dao.mapper;
-
-/**
- * @program: uncle-ray
- * @description:
- * @author: lei pei
- * @create: 2020-11-02 20:40
- */
-public interface StudentMapper {
-}
