--- src/main/java/com/uncleray/model/domain/Province.java
@@ -1,10 +1,0 @@
-package com.uncleray.model.domain;
-
-/**
- * @program: uncle-ray
- * @description:
- * @author: lei pei
- * @create: 2020-11-02 20:37
- */
-public class Province {
-}
