--- src/main/java/com/uncleray/model/entity/Student.java
@@ -1,10 +1,0 @@
-package com.uncleray.model.entity;
-
-/**
- * @program: uncle-ray
- * @description:
- * @author: lei pei
- * @create: 2020-11-02 20:36
- */
-public class Student {
-}
