--- src/main/java/com/uncleray/model/pojo/Area.java
@@ -1,24 +1,0 @@
-package com.uncleray.model.pojo;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Entity
-@Table
-public class Area {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer areaId;
-    private String  provinceId;
-    private String  provinceName;
-    private String address;
-    private String parentId;
-    private String parentName;
-
-}
