--- src/main/java/com/uncleray/model/pojo/Person.java
@@ -1,89 +1,0 @@
-package com.uncleray.model.pojo;
-
-import org.hibernate.annotations.Proxy;
-
-import javax.persistence.*;
-import java.io.Serializable;
-
-/**
- * @author uncle
- * @create 2020/7/17 - 23:37
- */
-@Entity
-@Proxy(lazy = false)
-@Table(name = "uncle_person")
-public class Person implements Serializable {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-
-    private String name;
-
-    private Integer age;
-
-    private char sex;
-
-    private String address;
-
-    private String graduation;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public Integer getAge() {
-        return age;
-    }
-
-    public void setAge(Integer age) {
-        this.age = age;
-    }
-
-    public char getSex() {
-        return sex;
-    }
-
-    public void setSex(char sex) {
-        this.sex = sex;
-    }
-
-    public String getAddress() {
-        return address;
-    }
-
-    public void setAddress(String address) {
-        this.address = address;
-    }
-
-    public String getGraduation() {
-        return graduation;
-    }
-
-    public void setGraduation(String graduation) {
-        this.graduation = graduation;
-    }
-
-    @Override
-    public String toString() {
-        return "Person{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", age=" + age +
-                ", sex=" + sex +
-                ", address='" + address + '\'' +
-                ", graduation='" + graduation + '\'' +
-                '}';
-    }
-}
