--- src/main/java/com/uncleray/service/AreaService.java
@@ -1,5 +1,0 @@
-package com.uncleray.service;
-
-public interface AreaService {
-    long countArea(String address);
-}
