+++ src/main/java/com/uncleray/util/SortUtils.java
@@ -1,8 +1,5 @@
 package com.uncleray.util;
 
-import com.uncleray.model.pojo.User;
-import org.springframework.util.Assert;
-
 /**
  * @program: uncle-ray
  * @description: 排序算法类
