+++ src/main/java/com/uncleray/web/dao/mapper/AreaMapper.java
@@ -1,0 +1,15 @@
+package com.uncleray.web.dao.mapper;
+
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+import com.uncleray.web.model.pojo.Area;
+import org.springframework.stereotype.Repository;
+
+/**
+ * @program: uncle-ray
+ * @description:
+ * @author: lei pei
+ * @create: 2020-11-02 20:40
+ */
+@Repository
+public interface AreaMapper extends BaseMapper<Area> {
+}
