+++ src/main/java/com/uncleray/web/model/domain/Province.java
@@ -1,0 +1,10 @@
+package com.uncleray.web.model.domain;
+
+/**
+ * @program: uncle-ray
+ * @description:
+ * @author: lei pei
+ * @create: 2020-11-02 20:37
+ */
+public class Province {
+}
