+++ src/main/java/com/uncleray/web/model/entity/Student.java
@@ -1,0 +1,10 @@
+package com.uncleray.web.model.entity;
+
+/**
+ * @program: uncle-ray
+ * @description:
+ * @author: lei pei
+ * @create: 2020-11-02 20:36
+ */
+public class Student {
+}
