+++ src/main/java/com/uncleray/web/service/AreaService.java
@@ -1,0 +1,5 @@
+package com.uncleray.web.service;
+
+public interface AreaService {
+    long countArea(String address);
+}
