+++ src/main/java/com/uncleray/web/service/impl/AreaServiceImpl.java
@@ -1,0 +1,13 @@
+package com.uncleray.web.service.impl;
+
+import com.uncleray.web.service.AreaService;
+import org.springframework.stereotype.Service;
+
+@Service
+public class AreaServiceImpl implements AreaService {
+
+    @Override
+    public long countArea(String address) {
+        return 0;
+    }
+}
