+++ src/main/resources/application-pro.yml
@@ -5,29 +5,18 @@ spring:
     name: uncle-springboot
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://127.0.0.1:3306/test?useUnicode=true&characterEncoding=utf8&serverTimezone=UTC
+    url: jdbc:mysql://127.0.0.1:3306/dev_uncle?useUnicode=true&characterEncoding=utf8&serverTimezone=UTC
     username: root
     password: 123456
-    hikari:
-      maximum-pool-size: 20
-  jpa:
-    database: mysql
-    show-sql: true
-    hibernate:
-      ddl-auto: update
-    properties:
-      hibernate:
-        format_sql: true
-        #配置数据的方言，这里是MySQL，还有比如Informix、Oracle、DB2等数据库
-        dialect: org.hibernate.dialect.MySQL5Dialect
   redis:
     host: 127.0.0.1
     port: 6379
-#        dialect: org.hibernate.dialect.MySQL5InnoDBDialect
-#        dialect: org.hibernate.dialect.MySQLMyISAMDialect
-#        enable_lazy_load_no_trans: true
-#        hbm2ddl:
-#          auto: create
+mybatis-plus:
+  configuration:
+    log-impl: org.apache.ibatis.logging.stdout.StdOutImpl
+  type-aliases-package: com.uncleray.model.pojo
+  mapper-locations: classpath:/mapper/*.xml
+
 springfox:
   documentation:
     swagger-ui:
