+++ src/main/resources/mapper/PersonMapper.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.uncleray.web.dao.mapper.PersonMapper">
+
+    <select id="selectList" resultType="com.uncleray.web.model.pojo.Person">
+        SELECT * FROM uncle_person WHERE age > 27
+    </select>
+</mapper>
