+++ src/test/java/com/uncleray/UncleRaysLoveApplicationTests.java
@@ -1,11 +1,9 @@
 package com.uncleray;
 
-import com.uncleray.model.pojo.Person;
+import com.uncleray.web.dao.mapper.PersonMapper;
+import com.uncleray.web.model.pojo.Person;
 import com.uncleray.result.ResultCodeEnum;
-import com.uncleray.dao.repository.PersonRepository;
-import com.uncleray.service.DemoService;
-import com.uncleray.service.MyService;
-import com.uncleray.service.PersonService;
+import com.uncleray.web.service.impl.PersonServiceImpl;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -13,7 +11,6 @@ import org.springframework.boot.test.context.SpringBootTest;
 import javax.sql.DataSource;
 import java.sql.Connection;
 import java.sql.SQLException;
-import java.util.List;
 
 @SpringBootTest
 class UncleRaysLoveApplicationTests {
@@ -30,64 +27,31 @@ class UncleRaysLoveApplicationTests {
     }
 
     @Autowired
-    PersonRepository repository;
+    PersonMapper personMapper;
     @Autowired
-    PersonService personService;
+    PersonServiceImpl personServiceImpl;
 
     @Test
     void myRepository() {
-        personService.deleteBatchPersons();
         Person person = new Person();
         person.setName("王妮");
         person.setAge(27);
         person.setSex('女');
         person.setGraduation("专科/无");
         person.setAddress("深圳市");
-        Person save = repository.save(person);
-        System.out.println(save);
-        personService.savePersons();
+        personMapper.insert(person);
+        ;
     }
-
-    @Test
-    void delete() {
-        personService.deleteBatchPersons();
-    }
-
-    @Autowired
-    DemoService demoService;
 
     @Test
     void test() {
-        List<Person> person = repository.testPerson("研究");
-        System.out.println(person);
-        int i = demoService.updatePerson(26, "雷佩");
-        System.out.println(i);
+
     }
 
-    @Autowired
-    MyService service;
 
     @Test
     void query() {
 
-//        Optional<Person> person = repository.findById(1);
-//        System.out.println(person.get());
-//
-//        long count = repository.count();
-//        System.out.println(count);
-//
-//        Person one = repository.getOne(2);
-//        System.out.println(one);
-
-        Person model = new Person();
-        model.setAge(26);
-        model.setName("雷佩");
-        model.setAddress("四川");
-        List<Person> search = service.findSearch(model);
-//        System.out.println(search);
-        for (Person person : search) {
-            System.out.println(person);
-        }
 
     }
